requirements.txt
aiodns==3.1.1
aiohttp==3.9.2
aiosignal==1.3.1
alpaca-py==0.14.0
alpha-vantage==2.3.1
altgraph @ file:///AppleInternal/Library/BuildRoots/1db8ea6b-8e8d-11ee-9041-cedaeb4cabe2/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
annotated-types==0.6.0
anyio==3.6.2
appdirs==1.4.4
APScheduler==3.10.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
bidict==0.22.1
blinker==1.7.0
ccxt==4.1.78
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
contourpy==1.2.0
cryptography==42.0.1
cycler==0.12.1
decorator==5.1.1
dnspython==2.5.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
exchange_calendars==4.5.1
executing==2.0.1
filelock==3.13.1
Flask==3.0.1
Flask-BabelEx==0.9.4
Flask-Login==0.6.3
Flask-Mail==0.9.1
flask-marshmallow==1.1.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.47.2
frozendict==2.4.0
frozenlist==1.4.1
fsspec==2023.12.2
future @ file:///AppleInternal/Library/BuildRoots/1db8ea6b-8e8d-11ee-9041-cedaeb4cabe2/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
h11==0.14.0
html5lib==1.1
huggingface-hub==0.20.3
humanfriendly==10.0
ibapi==9.81.1.post1
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
inflection==0.5.1
iniconfig==2.0.0
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jsonpickle==3.0.2
kiwisolver==1.4.5
korean-lunar-calendar==0.3.1
lumibot==3.0.5
lumiwealth-tradier==0.1.5
lxml==5.1.0
macholib @ file:///AppleInternal/Library/BuildRoots/1db8ea6b-8e8d-11ee-9041-cedaeb4cabe2/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
MarkupSafe==2.1.4
marshmallow==3.20.2
marshmallow-sqlalchemy==0.30.0
matplotlib==3.8.2
matplotlib-inline==0.1.6
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multitasking==0.0.11
networkx==3.2.1
numpy==1.26.3
packaging==23.2
pandas==2.0.3
pandas-datareader==0.10.0
pandas-market-calendars==4.3.2
parso==0.8.3
passlib==1.7.4
peewee==3.17.0
pexpect==4.9.0
pillow==10.2.0
plotly==5.18.0
pluggy==1.4.0
polygon-api-client==1.13.4
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pycares==4.4.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
pyluach==2.2.0
pyparsing==3.1.1
pytest==8.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.8.2
python-socketio==5.11.0
pytz==2023.4
PyYAML==6.0.1
Quandl==3.7.0
QuantStats==0.0.62
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scipy==1.10.1
seaborn==0.13.2
simple-websocket==1.0.0
six @ file:///AppleInternal/Library/BuildRoots/1db8ea6b-8e8d-11ee-9041-cedaeb4cabe2/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
sniffio==1.3.0
soupsieve==2.5
speaklater==1.3
SQLAlchemy==2.0.25
sseclient-py==1.8.0
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
timedelta==2020.12.3
tokenizers==0.15.1
tomli==2.0.1
toolz==0.12.1
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tqdm==4.66.1
traitlets==5.14.1
transformers==4.37.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==1.26.18
verboselogs==1.7
wcwidth==0.2.13
webencodings==0.5.1
websockets==11.0.3
Werkzeug==3.0.1
wsproto==1.2.0
WTForms==3.1.2
yarl==1.9.4
yfinance==0.2.36
zipp==3.17.0
